+++ DotNetDatenbankProtokollerV2/ProtokollerLibrary/Databases/MySQL/MySQLStorage.cs
@@ -63,9 +63,12 @@ namespace DotNetSimaticDatabaseProtokollerLibrary.Databases.MySQL
             }
         }
 
+        private DatasetConfig datasetConfig;
+
         public void CreateOrModify_TablesAndFields(string dataTable, DatasetConfig datasetConfig)
         {
             this.dataTable = dataTable;
+            this.datasetConfig = datasetConfig;
             this.fieldList = datasetConfig.DatasetConfigRows;
 
             string sql = "";
@@ -235,12 +238,14 @@ namespace DotNetSimaticDatabaseProtokollerLibrary.Databases.MySQL
 
                     if (_intValueList.Count > 0)
                     {
+                        bool ok = false;
+
                         lock (_intValueList)
                             _maxAdd = _intValueList.Count;
                         
                         try
                         {                           
-                            _internal_Write();                            
+                            ok = _internal_Write();                            
                         }
                         catch (ThreadAbortException)
                         {
@@ -254,7 +259,8 @@ namespace DotNetSimaticDatabaseProtokollerLibrary.Databases.MySQL
                                 Logging.LogText("Exception: ", ex, Logging.LogLevel.Error);
                         }
 
-                        _intValueList.RemoveRange(0, _maxAdd);
+                        if (ok)
+                            _intValueList.RemoveRange(0, _maxAdd);
                     }
                     else
                         Thread.Sleep(20);
@@ -282,19 +288,13 @@ namespace DotNetSimaticDatabaseProtokollerLibrary.Databases.MySQL
             catch (Exception)
             {
                 myDBConn.Close(); //Verbindung schließen!
+                myDBConn.Open();
                 if (myDBConn.State != System.Data.ConnectionState.Open)
                 {
-                    myDBConn.Open();
-                    if (myDBConn.State != System.Data.ConnectionState.Open)
-                    {
-                        return false;
-                    }
-                    myDBConn.ChangeDatabase(myConfig.Database);
-                }
-                else
-                {
+                    Logging.LogText("Error ReConnecting to Database! Dataset:" + datasetConfig.Name, Logging.LogLevel.Error);
                     return false;
                 }
+                myDBConn.ChangeDatabase(myConfig.Database);
             }
 
             //Add the Fields to the Database
