+++ DotNetDatenbankProtokollerV2/ProtokollerLibrary/Databases/PostgreSQL/PostgreSQLStorage.cs
@@ -230,16 +230,16 @@ namespace DotNetSimaticDatabaseProtokollerLibrary.Databases.PostgreSQL
             {
                 while (true)
                 {
-
-
                     if (_intValueList.Count > 0)
                     {
+                        bool ok = false;
+
                         lock (_intValueList)
                             _maxAdd = _intValueList.Count;
                         
                         try
                         {                           
-                            _internal_Write();                            
+                            ok = _internal_Write();                            
                         }
                         catch (ThreadAbortException)
                         {
@@ -253,7 +253,8 @@ namespace DotNetSimaticDatabaseProtokollerLibrary.Databases.PostgreSQL
                                 Logging.LogText("Exception: ", ex, Logging.LogLevel.Error);
                         }
 
-                        _intValueList.RemoveRange(0, _maxAdd);
+                        if (ok)
+                            _intValueList.RemoveRange(0, _maxAdd);
                     }
                     else
                         Thread.Sleep(20);
@@ -281,19 +282,13 @@ namespace DotNetSimaticDatabaseProtokollerLibrary.Databases.PostgreSQL
             catch (Exception)
             {
                 myDBConn.Close(); //Verbindung schließen!
+                myDBConn.Open();
                 if (myDBConn.State != System.Data.ConnectionState.Open)
                 {
-                    myDBConn.Open();
-                    if (myDBConn.State != System.Data.ConnectionState.Open)
-                    {
-                        return false;
-                    }
-                    myDBConn.ChangeDatabase(myConfig.Database);
-                }
-                else
-                {
+                    Logging.LogText("Error ReConnecting to Database! Dataset:" + datasetConfig.Name, Logging.LogLevel.Error);
                     return false;
                 }
+                myDBConn.ChangeDatabase(myConfig.Database);
             }
 
             //Add the Fields to the Database
