+++ DotNetDatenbankProtokollerV2/ProtokollerLibrary/Databases/SQLite/SQLLiteStorage.cs
@@ -242,12 +242,13 @@ namespace DotNetSimaticDatabaseProtokollerLibrary.Databases.SQLite
 
                     if (_intValueList.Count > 0)
                     {
+                        bool ok = false;
                         lock (_intValueList)
                             _maxAdd = _intValueList.Count;
                         
                         try
                         {                           
-                            _internal_Write();                            
+                            ok = _internal_Write();                            
                         }
                         catch (ThreadAbortException)
                         {
@@ -261,7 +262,8 @@ namespace DotNetSimaticDatabaseProtokollerLibrary.Databases.SQLite
                                 Logging.LogText("Exception: ", ex, Logging.LogLevel.Error);
                         }
 
-                        _intValueList.RemoveRange(0, _maxAdd);
+                        if (ok)
+                            _intValueList.RemoveRange(0, _maxAdd);
                     }
                     else
                         Thread.Sleep(20);
@@ -291,16 +293,10 @@ namespace DotNetSimaticDatabaseProtokollerLibrary.Databases.SQLite
                 catch (Exception)
                 {
                     myDBConn.Close(); //Verbindung schließen!
+                    myDBConn.Open();
                     if (myDBConn.State != System.Data.ConnectionState.Open)
                     {
-                        myDBConn.Open();
-                        if (myDBConn.State != System.Data.ConnectionState.Open)
-                        {
-                            return false;
-                        }
-                    }
-                    else
-                    {
+                        Logging.LogText("Error ReConnecting to Database! Dataset:" + datasetConfig.Name, Logging.LogLevel.Error);
                         return false;
                     }
                 }
